% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_pressure.R
\name{convert_pressure}
\alias{convert_pressure}
\title{Pressure Unit Conversion}
\usage{
convert_pressure(value, fromUnit = c("Pa", "bar", "atm", "Torr", "mmHg"),
  toUnit = c("Pa", "bar", "atm", "Torr", "mmHg"), dp = 2)
}
\arguments{
\item{value}{Number to be converted}

\item{fromUnit}{Unit to be converted from "Pa", "bar", "atm", "Torr" or "mmHg"}

\item{toUnit}{Unit to be converted to "Pa", "bar", "atm", "Torr" or "mmHg"}

\item{dp}{Number of decimal places to be rounded to}
}
\value{
Value of converted Unit
}
\description{
Pressure Unit Conversion
}
