% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_weight.R
\name{convert_weight}
\alias{convert_weight}
\title{Weight Unit Conversion}
\usage{
convert_weight(value, fromUnit = c("g", "kg", "gr", "oz", "lb"),
  toUnit = c("g", "kg", "gr", "oz", "lb"), dp = 2)
}
\arguments{
\item{value}{Number to be converted}

\item{fromUnit}{Unit to be converted from "g", "kg", "gr", "oz" or "lb"}

\item{toUnit}{Unit to be converted to "g", "kg", "gr", "oz" or "lb"}

\item{dp}{Number of decimal places to be rounded to}
}
\value{
Value of converted Unit
}
\description{
Weight Unit Conversion
}
